requirements.txt
aiohttp==3.4.4
ansimarkup==1.4.0
astroid==2.1.0
async-timeout==3.0.1
attrs==18.2.0
better-exceptions-fork==0.2.1.post6
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.0.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
loguru==0.2.1
MarkupSafe==1.1.0
mccabe==0.6.1
multidict==4.5.2
Pygments==2.3.0
python-dotenv==0.10.0
python-wowapi==2.1.0
requests==2.20.1
six==1.12.0
telepot==12.7
urllib3==1.24.1
waitress==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
yarl==1.2.6
